apnsworker: move payload package into models package
